Move non-studio-specific hooks to @foxglove/hooks package
