+++ packages/@types/foxglove__web/index.d.ts
