+++ packages/@types/foxglove__web/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@types/foxglove__web",
+  "types": "index.d.ts"
+}
