+++ packages/hooks/src/index.ts
@@ -2,9 +2,17 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
+export { default as useChangeDetector } from "./useChangeDetector";
+export { default as useDeepMemo } from "./useDeepMemo";
+export { default as useGuaranteedContext } from "./useGuaranteedContext";
+export { default as useMustNotChange } from "./useMustNotChange";
 export { default as useRethrow } from "./useRethrow";
 export { default as useShallowMemo } from "./useShallowMemo";
-export { default as useVisibilityState } from "./useVisibilityState";
-export { default as useMustNotChange } from "./useMustNotChange";
 export { default as useValueChangedDebugLog } from "./useValueChangedDebugLog";
+export { default as useVisibilityState } from "./useVisibilityState";
+export { default as useWarnImmediateReRender } from "./useWarnImmediateReRender";
+export { useMemoryInfo } from "./useMemoryInfo";
+export * from "./selectWithUnstableIdentityWarning";
 export * from "./useCrash";
+export * from "./useSessionStorageValue";
+export * from "./useSynchronousMountedState";
