+++ packages/hooks/src/useDeepMemo.test.ts
@@ -14,7 +14,7 @@
 
 import { renderHook } from "@testing-library/react-hooks";
 
-import useDeepMemo from "@foxglove/studio-base/hooks/useDeepMemo";
+import useDeepMemo from "./useDeepMemo";
 
 describe("useDeepMemo", () => {
   it("returns original object when deep equal", () => {
