+++ packages/hooks/src/useMemoryInfo.ts
@@ -12,20 +12,18 @@ type UseMemoryInfoOptions = {
   refreshIntervalMs: number;
 };
 
-const performance = window.performance;
-
 export function useMemoryInfo(opt: UseMemoryInfoOptions): MemoryInfo | undefined {
   const { refreshIntervalMs } = opt;
-  const [memoryInfo, setMemoryInfo] = useState<MemoryInfo | undefined>(performance.memory);
+  const [memoryInfo, setMemoryInfo] = useState<MemoryInfo | undefined>(window.performance.memory);
 
   useEffect(() => {
-    if (!performance.memory) {
+    if (!window.performance.memory) {
       log.info("No memory information available");
       return;
     }
 
     const interval = setInterval(() => {
-      setMemoryInfo(performance.memory);
+      setMemoryInfo(window.performance.memory);
     }, refreshIntervalMs);
     return () => {
       clearInterval(interval);
