+++ packages/studio-base/package.json
@@ -61,6 +61,7 @@
     "@types/base16": "^1.0.2",
     "@types/cytoscape": "^3.19.9",
     "@types/foxglove__roslibjs": "workspace:*",
+    "@types/foxglove__web": "workspace:*",
     "@types/gl-matrix": "3.2.0",
     "@types/hammerjs": "2.0.41",
     "@types/leaflet": "1.9.3",
