+++ packages/studio-base/src/context/TimelineInteractionStateContext.tsx
@@ -15,8 +15,8 @@ import { createContext, useCallback } from "react";
 import { DeepReadonly } from "ts-essentials";
 import { StoreApi, useStore } from "zustand";
 
+import { useGuaranteedContext } from "@foxglove/hooks";
 import { TimelinePositionedEvent } from "@foxglove/studio-base/context/EventsContext";
-import useGuaranteedContext from "@foxglove/studio-base/hooks/useGuaranteedContext";
 import type { HoverValue } from "@foxglove/studio-base/types/hoverValue";
 
 /**
