+++ packages/studio-base/src/context/Workspace/WorkspaceContext.ts
@@ -5,10 +5,10 @@
 import { createContext } from "react";
 import { StoreApi, useStore } from "zustand";
 
+import { useGuaranteedContext } from "@foxglove/hooks";
 import { AppSettingsTab } from "@foxglove/studio-base/components/AppSettingsDialog/AppSettingsDialog";
 import { DataSourceDialogItem } from "@foxglove/studio-base/components/DataSourceDialog";
 import { IDataSourceFactory } from "@foxglove/studio-base/context/PlayerSelectionContext";
-import useGuaranteedContext from "@foxglove/studio-base/hooks/useGuaranteedContext";
 
 export const SidebarItemKeys = [
   "account",
