+++ packages/studio-base/src/hooks/index.ts
@@ -6,4 +6,3 @@
 
 export { useAppConfigurationValue } from "./useAppConfigurationValue";
 export { useAppTimeFormat } from "./useAppTimeFormat";
-export * from "./useMemoryInfo";
