+++ packages/studio-base/src/typings/index.d.ts
@@ -7,6 +7,6 @@
 import "./extensions";
 import "./react";
 import "./overrides";
-import "./web";
+import "../../../@types/web/ind";
 import "./webpack-defines";
 import "./i18next";
